Fixed Bug Issue - shuffle delay - undefined index in an array - issue in fisher yates sorting
