+++ index.htm
@@ -44,6 +44,6 @@
     </div>
     
 
-    <script src="script.js" defer></script>
+    <script type="module" src="script.js"></script>
 </body>
 </html>
